# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Passes for Private Residential Development",
    "description": "API to manage access passes for a gated residential community, including pass review, QR generation and gatehouse validation.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/passes": {
      "post": {
        "tags": [
          "Passes"
        ],
        "summary": "Create Simple Pass",
        "operationId": "create_simple_pass_passes_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePassesSimple"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassCreatedResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "House user not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Passes"
        ],
        "summary": "Get Passes For User",
        "operationId": "get_passes_for_user_passes_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassesResponseList"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/passes/days": {
      "post": {
        "tags": [
          "Passes"
        ],
        "summary": "Create Pass For Days",
        "operationId": "create_pass_for_days_passes_days_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePassesForDays"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassCreatedForDaysResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/passes/{pass_id}/qr": {
      "get": {
        "tags": [
          "Passes"
        ],
        "summary": "Get Pass Qr",
        "operationId": "get_pass_qr_passes__pass_id__qr_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pass_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pass Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassQRCodeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/passes/all": {
      "get": {
        "tags": [
          "Passes"
        ],
        "summary": "Get All Passes",
        "operationId": "get_all_passes_passes_all_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "next_cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Next Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 15,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassesPageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/passes/search": {
      "get": {
        "tags": [
          "Passes"
        ],
        "summary": "Search Pass By Id",
        "operationId": "search_pass_by_id_passes_search_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pass_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pass Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Search Pass By Id Passes Search Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/passes/pending/count": {
      "get": {
        "tags": [
          "Passes"
        ],
        "summary": "Count Pending Passes",
        "operationId": "count_pending_passes_passes_pending_count_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingPassesCountResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/passes/pending": {
      "get": {
        "tags": [
          "Passes"
        ],
        "summary": "Get Pending Passes",
        "operationId": "get_pending_passes_passes_pending_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PassesResponseUser"
                  },
                  "title": "Response Get Pending Passes Passes Pending Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/passes/{pass_id}/approve": {
      "patch": {
        "tags": [
          "Passes"
        ],
        "summary": "Approve Pass",
        "operationId": "approve_pass_passes__pass_id__approve_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pass_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pass Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/passes/{pass_id}/reject": {
      "post": {
        "tags": [
          "Passes"
        ],
        "summary": "Reject a pending pass",
        "operationId": "reject_pass_passes__pass_id__reject_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pass_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pass Id"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reason"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Passes"
        ],
        "summary": "Reject a pending pass (deprecated, use POST)",
        "operationId": "reject_pass_deprecated_passes__pass_id__reject_delete",
        "deprecated": true,
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pass_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pass Id"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reason"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/passes/{pass_id}/auto-review": {
      "patch": {
        "tags": [
          "Passes"
        ],
        "summary": "Auto Review Pass",
        "operationId": "auto_review_pass_passes__pass_id__auto_review_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pass_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pass Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewSchema"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePassesForDays": {
        "properties": {
          "days": {
            "type": "integer",
            "exclusiveMaximum": 8.0,
            "exclusiveMinimum": 0.0,
            "title": "Days"
          },
          "guest_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Guest Name"
          },
          "valid_from": {
            "type": "string",
            "format": "date-time",
            "title": "Valid From"
          },
          "reason": {
            "type": "string",
            "maxLength": 300,
            "minLength": 10,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "days",
          "guest_name",
          "valid_from",
          "reason"
        ],
        "title": "CreatePassesForDays"
      },
      "CreatePassesSimple": {
        "properties": {
          "pass_type": {
            "$ref": "#/components/schemas/PassTypeSimple"
          },
          "guest_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Guest Name"
          },
          "valid_from": {
            "type": "string",
            "format": "date-time",
            "title": "Valid From"
          }
        },
        "type": "object",
        "required": [
          "pass_type",
          "guest_name",
          "valid_from"
        ],
        "title": "CreatePassesSimple"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse"
      },
      "PassCreatedForDaysResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "pass_id": {
            "type": "string",
            "title": "Pass Id"
          }
        },
        "type": "object",
        "required": [
          "message",
          "pass_id"
        ],
        "title": "PassCreatedForDaysResponse"
      },
      "PassCreatedResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "pass": {
            "additionalProperties": true,
            "type": "object",
            "title": "Pass"
          }
        },
        "type": "object",
        "required": [
          "message",
          "pass"
        ],
        "title": "PassCreatedResponse"
      },
      "PassQRCodeResponse": {
        "properties": {
          "qr_jpg_code_base64": {
            "type": "string",
            "title": "Qr Jpg Code Base64"
          }
        },
        "type": "object",
        "required": [
          "qr_jpg_code_base64"
        ],
        "title": "PassQRCodeResponse"
      },
      "PassTypeSimple": {
        "type": "string",
        "enum": [
          "temporary",
          "temporary_party",
          "temporary_gym"
        ],
        "title": "PassTypeSimple"
      },
      "PassesPageResponse": {
        "properties": {
          "passes": {
            "items": {
              "$ref": "#/components/schemas/PassesResponseUser"
            },
            "type": "array",
            "title": "Passes"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          }
        },
        "type": "object",
        "required": [
          "passes",
          "has_next"
        ],
        "title": "PassesPageResponse"
      },
      "PassesResponseList": {
        "properties": {
          "passes": {
            "items": {
              "$ref": "#/components/schemas/PassesResponseUser"
            },
            "type": "array",
            "title": "Passes"
          }
        },
        "type": "object",
        "required": [
          "passes"
        ],
        "title": "PassesResponseList"
      },
      "PassesResponseUser": {
        "properties": {
          "_id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "valid_from": {
            "type": "string",
            "format": "date-time",
            "title": "Valid From"
          },
          "valid_until": {
            "type": "string",
            "format": "date-time",
            "title": "Valid Until"
          }
        },
        "type": "object",
        "required": [
          "_id",
          "status",
          "enabled",
          "valid_from",
          "valid_until"
        ],
        "title": "PassesResponseUser"
      },
...
      "PendingPassesCountResponse": {
        "properties": {
          "pending_passes": {
            "type": "integer",
            "title": "Pending Passes"
          }
        },
        "type": "object",
        "required": [
          "pending_passes"
        ],
        "title": "PendingPassesCountResponse"
      },
      "ReviewSchema": {
        "properties": {
          "approved": {
            "type": "boolean",
            "title": "Approved"
          },
          "reason": {
            "type": "string",
            "maxLength": 300,
            "minLength": 10,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "approved",
          "reason"
        ],
        "title": "ReviewSchema"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}